fix: документированный путь до объекта в Board теперь действительно собирается
Ревью нашло, что образец новой архитектуры был недостижим из cross, а его
код не компилировался. Три отдельные поломки:

adapters не был объявлен ни в crates-cross/Cargo.toml, ни у bsp — то есть
пользователь, следующий инструкции из resources.rs, получал E0433 на первой
же строке. Теперь крейт объявлен в обоих местах, и путь работает без правки
манифестов.

adapters::Lm75::<_>::DEFAULT_ADDRESS не выводится (E0282): константа сидела в
impl<I2C> Lm75<I2C>, хотя от шины не зависит, и каждый вызов требовал бы
турбофиша с конкретным типом шины. Вынесена в свободную adapters::lm75::
DEFAULT_ADDRESS — тесты заодно перестали писать Lm75::<Mock>::.

mod lm75 был приватным, и обоснование выбора датчика вместе с разбором
формата не попадало в cargo doc — ровно от тех, кому адресовано. Теперь
pub mod, реэкспорт Lm75 сохранён.

Заодно две вещи из того же ревью:

FirmwareUpdate::write брал usize, а read — u32; порт копировал асимметрию
embassy-boot. На 32-битной цели незаметно, зато у фейка под хост одно и то же
смещение перестало бы выражаться обоими типами. Оба теперь u32, bsp приводит
на границе.

domain не зависел от ports, хотя доки обоих крейтов обещают, что порты —
граница домена и ими пользуется логика. Зависимость добавлена: иначе правило
«три неудачных старта — откатиться» нельзя было бы даже написать в domain.

Проверено: template-check по девяти конфигурациям и прогон на STM32F303VC —
5 target-тестов и 2 host-target.

+++ crates-cross/bsp/src/ota.rs
@@ -198,8 +198,11 @@ impl ports::FirmwareUpdate for Ota {
     /// Стиранием секторов `embassy-boot` занимается сам, но длина куска должна
     /// быть кратна `WRITE_SIZE` флеша — обычно это и есть размер пакета,
     /// которым транспорт отдаёт данные.
-    fn write(&mut self, offset: usize, data: &[u8]) -> Result<(), Self::Error> {
-        self.updater().write_firmware(offset, data)
+    fn write(&mut self, offset: u32, data: &[u8]) -> Result<(), Self::Error> {
+        // `write_firmware` берёт `usize`, а порт — `u32` (см. его
+        // doc-комментарий): на этой цели типы совпадают по ширине, но порт
+        // должен одинаково выражаться и на хосте, где живут фейки.
+        self.updater().write_firmware(offset as usize, data)
     }
 
     /// Читает записанное обратно — проверить контрольную сумму принятого
